ref: merge model with interface
