+++ src/models/index.ts
@@ -2,6 +2,5 @@ export * from './application-identifier';
 export * from './application-identifiers';
 export * from './barcode-value';
 export * from './embedded-data';
-export * from './ibarcode-value';
 export * from './parser.configuration';
 export * from './reader.configuration';
